refactor(gallery): extract logic from view to separate component
