+++ src/components/gallery/gallery.component.html
@@ -1,0 +1,1 @@
+<mdlv-photo-swipe [images]="filteredImages"></mdlv-photo-swipe>
