+++ src/components/index.ts
@@ -6,4 +6,4 @@ export * from './contact-form/contact-form.component';
 export * from './pricing-form/pricing-form.component';
 export * from './offer-presentation/offer-presentation.component';
 export * from './offer-banner/offer-banner.component';
-export * from './gallery/photo-swipe/photo-swipe.component';
+export * from './gallery/gallery.component';
