+++ src/views/gallery-page/gallery-page.component.html
@@ -18,4 +18,4 @@
 
 <mat-tab-nav-panel #tabPanel></mat-tab-nav-panel>
 
-<mdlv-photo-swipe [images]="filteredImages"></mdlv-photo-swipe>
+<mdlv-gallery [activeCategory]="activeLink.label"></mdlv-gallery>
