+++ src/views/gallery-page/gallery-page.component.scss
@@ -1,24 +1,3 @@
 mat-tab-group {
   margin-bottom: 16px;
 }
-
-.gallery-container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  padding-bottom: 80px;
-}
-
-.gallery-item {
-  margin: 5px;
-}
-
-.spinner-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: fixed;
-  bottom: 0;
-  left: 50%;
-  margin-bottom: 16px;
-}
