+++ src/views/gallery-page/gallery-page.component.ts
@@ -1,25 +1,15 @@
 import { ActivatedRoute, Router, RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
-import { Component, HostListener } from '@angular/core';
-import { MatGridListModule } from '@angular/material/grid-list';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+import { Component } from '@angular/core';
 import { MatTabsModule } from '@angular/material/tabs';
 
-import { GalleryImage } from '@models/gallery.model';
-import { PhotoSwipeComponent } from '@components/index';
+import { GalleryComponent } from '@components/index';
 
 type Link = { label: string; route: string };
 
 @Component({
   selector: 'mdlv-gallery-page',
-  imports: [
-    CommonModule,
-    MatGridListModule,
-    MatTabsModule,
-    MatProgressSpinnerModule,
-    RouterModule,
-    PhotoSwipeComponent,
-  ],
+  imports: [CommonModule, MatTabsModule, RouterModule, GalleryComponent],
   templateUrl: './gallery-page.component.html',
   styleUrl: './gallery-page.component.scss',
 })
@@ -48,16 +38,10 @@ export class GalleryPageComponent {
   ];
   activeLink!: Link;
 
-  images: GalleryImage[] = [];
-  isLoading: boolean = false;
-
   constructor(private route: ActivatedRoute, private router: Router) {}
 
   ngOnInit(): void {
     this.activeLink = this.getActiveLink();
-
-    this.images = this.loadImages(50);
-    this.refreshLightGallery();
   }
 
   getActiveLink(): Link {
@@ -68,56 +52,5 @@ export class GalleryPageComponent {
     // Find the matching link based on the last segment.
     const foundLink = this.links.find((link) => link.route === lastSegment);
     return foundLink ? foundLink : this.links[0];
-  }
-
-  @HostListener('window:scroll')
-  onScroll(): void {
-    const scrollPosition = window.scrollY + window.innerHeight;
-    const documentHeight = document.documentElement.scrollHeight;
-
-    if (scrollPosition >= documentHeight - 10 && !this.isLoading) {
-      this.isLoading = true;
-      setTimeout(() => {
-        this.images = [...this.images, ...this.loadImages(20)];
-        this.refreshLightGallery();
-        this.isLoading = false;
-      }, 1000);
-    }
-  }
-
-  loadImages(count: number): GalleryImage[] {
-    return Array.from({ length: count }, (_, i) => {
-      const index = this.images.length + i;
-      return {
-        title: `Event ${index}`,
-        imageUri: `https://picsum.photos/2000/1500?random=${index}`,
-        imageSize: '4000-3000',
-        thumbnailUri: `https://picsum.photos/200/150?random=${index}`,
-        category: this.getRandomCategory(),
-      };
-    });
-  }
-
-  refreshLightGallery(): void {}
-
-  // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-  getRandomCategory(): any {
-    // temporary function to generate random category
-    const categories: string[] = [
-      'imprezy firmowe',
-      'imprezy prywatne',
-      'imprezy plenerowe',
-      'prezentacje',
-    ];
-    const randomIndex = Math.floor(Math.random() * categories.length);
-    return categories[randomIndex];
-  }
-
-  get filteredImages(): GalleryImage[] {
-    return this.images.filter(
-      (image) =>
-        this.activeLink.label === 'wszystkie' ||
-        image.category === this.activeLink.label
-    );
   }
 }
